Add separate function for URL background fetching
Converted the command string that was passed to the python interpreter
into a separate function and built a wrapper around it that reads its
source via the inspect module, prepares it for execution and passes it
to the interpreter as before (passing the needed arguments).

While this is a little more complex it gets us the convenience of the
interpreter checking syntax before actually executing the script inside
of WeeChat and thus making the whole thing much more maintainable.

The function fetchit() also has some added functionality. First of all
exceptions are now properly handled. Secondly the input is now read in
1024 until the read data includes "</title>".